Fix
Some changes in search method, to calculate medan correctly and to round the numbr to second digit. Change in sorting of data, to get sorted date directly from DB and not to sort them separatly. Change of DBInit class, to save only stations that are open (this was something that i miss).